add tabLayout in RankingFragment
