+++ app/src/main/res/layout/fragment_book_ranking.xml
@@ -1,0 +1,146 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".Fragments.BookRankingFragment">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_gravity="center"
+        android:layout_margin="10dp"
+        android:background="@color/white"
+        android:gravity="center"
+        android:orientation="vertical">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginTop="30dp"
+                android:orientation="vertical">
+
+                <TextView
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="top"
+                    android:fontFamily="@font/font_main"
+                    android:gravity="center"
+                    android:text="회원들이 직접 뽑은"
+                    android:textStyle="bold" />
+
+                <TextView
+                    android:id="@+id/tv_rankingtag"
+                    android:layout_width="match_parent"
+                    android:layout_height="wrap_content"
+                    android:layout_gravity="top"
+                    android:layout_marginTop="10dp"
+                    android:fontFamily="@font/font_main"
+                    android:gravity="center"
+                    android:text="가장 많이 담긴 도서"
+                    android:textSize="25sp"
+                    android:textStyle="bold" />
+
+            </LinearLayout>
+
+            <RelativeLayout
+                android:layout_width="match_parent"
+                android:layout_height="180dp"
+                android:layout_marginTop="10dp"
+                android:layout_marginBottom="10dp"
+                android:gravity="center">
+
+                <TextView
+                    android:id="@+id/tv_1st"
+                    android:layout_width="50dp"
+                    android:layout_height="25dp"
+                    android:layout_above="@+id/iv_first"
+                    android:background="@drawable/hlafcircle"
+                    android:fontFamily="@font/font_main"
+                    android:gravity="center"
+                    android:text="1등 "
+                    android:textColor="@color/white"
+                    android:textStyle="bold" />
+
+                <ImageView
+                    android:id="@+id/iv_first"
+                    android:layout_width="85dp"
+                    android:layout_height="110dp"
+                    android:layout_centerVertical="true"
+                    android:layout_marginRight="15dp"
+                    android:background="@color/white"
+                    android:elevation="10dp" />
+
+                <TextView
+                    android:id="@+id/tv_2nd"
+                    android:layout_width="50dp"
+                    android:layout_height="25dp"
+                    android:layout_above="@id/iv_second"
+                    android:layout_toRightOf="@id/iv_first"
+                    android:background="@drawable/hlafcircle"
+                    android:fontFamily="@font/font_main"
+                    android:gravity="center"
+                    android:text="2등 "
+                    android:textColor="@color/white"
+                    android:textStyle="bold" />
+
+                <ImageView
+                    android:id="@+id/iv_second"
+                    android:layout_width="85dp"
+                    android:layout_height="110dp"
+                    android:layout_centerVertical="true"
+                    android:layout_marginRight="15dp"
+                    android:layout_toRightOf="@id/iv_first"
+                    android:background="@color/white"
+                    android:elevation="10dp" />
+
+                <TextView
+                    android:id="@+id/tv_3rd"
+                    android:layout_width="50dp"
+                    android:layout_height="25dp"
+                    android:layout_above="@id/iv_third"
+                    android:layout_toRightOf="@id/iv_second"
+                    android:background="@drawable/hlafcircle"
+                    android:fontFamily="@font/font_main"
+                    android:gravity="center"
+                    android:text="3등 "
+                    android:textColor="@color/white"
+                    android:textStyle="bold" />
+
+                <ImageView
+                    android:id="@+id/iv_third"
+                    android:layout_width="85dp"
+                    android:layout_height="110dp"
+                    android:layout_centerVertical="true"
+                    android:layout_toRightOf="@id/iv_second"
+                    android:background="@color/white"
+                    android:elevation="10dp" />
+
+            </RelativeLayout>
+
+        </LinearLayout>
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_marginTop="15dp"
+            android:background="@color/whiteCardColor"
+            android:gravity="center">
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/rv_ranking"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center"
+                android:scrollbarFadeDuration="0"
+                android:theme="@style/edgeEffect" />
+
+        </LinearLayout>
+    </LinearLayout>
+</LinearLayout>
