+++ app/src/main/res/menu/menu_bottomnavigationview.xml
@@ -16,7 +16,7 @@
     <item
         android:id="@+id/chatting"
         android:icon="@drawable/icon_boycrown"
-        android:title="독서짱" />
+        android:title="고민중" />
     <item
         android:id="@+id/myPage"
         android:icon="@drawable/icon_user"
